cs 3.8 Code cleanup

+++ SpicesUpdate@claudiux/files/SpicesUpdate@claudiux/cs/3.8/modules/cs_themes.py
@@ -4,7 +4,7 @@ from gi.repository.Gtk import SizeGroup, SizeGroupMode
 
 from GSettingsWidgets import *
 from ExtensionCore import DownloadSpicesPage
-from SUSpices import Spice_Harvester
+from SUSpices import SU_Spice_Harvester
 
 import glob
 
@@ -27,7 +27,7 @@ class Module:
         if not self.loaded:
             print("Loading Themes module")
 
-            self.spices = Spice_Harvester('theme', self.window)
+            self.spices = SU_Spice_Harvester('theme', self.window)
 
             self.sidePage.stack = SettingsStack()
             self.sidePage.add_widget(self.sidePage.stack)
@@ -149,16 +149,18 @@ class Module:
                         if os.path.exists(path):
                             chooser.add_picture(path, callback, title=theme_name, id=theme_name)
                             break
-                except:
+                except Exception:
                     chooser.add_picture("/usr/share/cinnamon/thumbnails/%s/unknown.png" % path_suffix, callback, title=theme_name, id=theme_name)
                 GLib.timeout_add(5, self.increment_progress, (chooser, inc))
         GLib.timeout_add(500, self.hide_progress, chooser)
 
     def increment_progress(self, payload):
+        if self is None: return
         (chooser, inc) = payload
         chooser.increment_loading_progress(inc)
 
     def hide_progress(self, chooser):
+        if self is None: return
         chooser.set_sensitive(True)
         chooser.reset_loading_progress()
 
@@ -166,7 +168,7 @@ class Module:
         self.window = window
 
     def make_group(self, group_label, widget, add_widget_to_size_group=True):
-        self.size_groups = getattr(self, "size_groups", [Gtk.SizeGroup.new(Gtk.SizeGroupMode.HORIZONTAL) for x in range(2)])
+        self.size_groups = getattr(self, "size_groups", [SizeGroup.new(SizeGroupMode.HORIZONTAL) for x in range(2)])
         box = SettingsWidget()
         label = Gtk.Label()
         label.set_markup(group_label)
@@ -180,6 +182,7 @@ class Module:
         return box
 
     def create_button_chooser(self, settings, key, path_prefix, path_suffix, button_picture_size, menu_pictures_size, num_cols):
+        if self is None: return
         chooser = PictureChooserButton(num_cols=num_cols, button_picture_size=button_picture_size, menu_pictures_size=menu_pictures_size, has_button_label=True)
         theme = settings.get_string(key)
         chooser.set_button_label(theme)
@@ -201,7 +204,7 @@ class Module:
                     if os.path.exists(path):
                         chooser.set_picture_from_file(path)
                         break
-            except:
+            except Exception:
                 chooser.set_picture_from_file("/usr/share/cinnamon/thumbnails/%s/unknown.png" % path_suffix)
         return chooser
 
@@ -269,16 +272,18 @@ class Module:
         return res
 
     def filter_func_gtk_dir(self, directory):
+        if self is None: return
         # returns whether a directory is a valid GTK theme
         if os.path.exists(os.path.join(directory, "gtk-2.0")):
             if os.path.exists(os.path.join(directory, "gtk-3.0")):
                 return True
             else:
-                for subdir in glob.glob("%s/gtk-3.*" % directory):
+                if len(glob.glob("%s/gtk-3.*" % directory)) > 0:
                     return True
         return False
 
     def _load_icon_themes(self):
+        if self is None: return
         dirs = ("/usr/share/icons", os.path.join(os.path.expanduser("~"), ".icons"))
         walked = walk_directories(dirs, lambda d: os.path.isdir(d), return_directories=True)
         valid = []
@@ -306,6 +311,7 @@ class Module:
         return res
 
     def _load_cursor_themes(self):
+        if self is None: return
         dirs = ("/usr/share/icons", os.path.join(os.path.expanduser("~"), ".icons"))
         valid = walk_directories(dirs, lambda d: os.path.isdir(d) and os.path.exists(os.path.join(d, "cursors")), return_directories=True)
         valid.sort(key=lambda a: a[0].lower())
@@ -321,6 +327,7 @@ class Module:
         return res
 
     def _load_metacity_themes(self):
+        if self is None: return
         dirs = ("/usr/share/themes", os.path.join(os.path.expanduser("~"), ".themes"))
         valid = walk_directories(dirs, lambda d: os.path.exists(os.path.join(d, "metacity-1/metacity-theme-3.xml")), return_directories=True)
         valid.sort(key=lambda a: a[0].lower())
@@ -336,6 +343,7 @@ class Module:
         return res
 
     def _load_cinnamon_themes(self):
+        if self is None: return
         dirs = ("/usr/share/themes", os.path.join(os.path.expanduser("~"), ".themes"))
         valid = walk_directories(dirs, lambda d: os.path.exists(os.path.join(d, "cinnamon")), return_directories=True)
         valid.sort(key=lambda a: a[0].lower())
@@ -351,13 +359,14 @@ class Module:
         return res
 
     def update_cursor_theme_link(self, path, name):
+        if self is None: return
         default_dir = os.path.join(os.path.expanduser("~"), ".icons", "default")
         index_path = os.path.join(default_dir, "index.theme")
 
         try:
             os.makedirs(default_dir)
         except os.error as e:
-            pass
+            print(e)
 
         if os.path.exists(index_path):
             os.unlink(index_path)
